requirements.txt
absl-py
certifi
charset-normalizer
contextlib2
filelock
fsspec
idna
iniconfig
Jinja2
MarkupSafe
ml-collections
mpmath
networkx
numpy
packaging
Pillow
pluggy
pygame
pytest
PyYAML
requests
six
sympy
torch
torchaudio
torchvision
typing_extensions
urllib3
